Reorganization and map changes
This commit does 3 things:
- it moves any component in the frontend to the
relevant subfolder to make the items easier to
navigate
- it moved the MapComponent constants out of the
main function. This was done based on codeRabbit comment
pointing out that this forced re-running the map render
and updating the state every time another state was changed
which could lead to constant re-renders. Thus the move out
of the main function should avoid this.
- it fixes a bunch of small HTML changes to avoir invalid
characters showing up in the popups as well as deletes
some unused test variables which were still hanging around
the code base

GraphiqueAnaVar
Moved from client/src/components/ to client/src/components/charts
added a space to one  of the lablels

carteEnsRegTerr
Moved from client/src/components to client/src/components/maps
Moved MapComponent out of CarteEnsRegTerr main function and
manipulated the varaiables as required.
Added an escape function that converts characters to their HTML
representation

carteHistorique
Moved from client/src/components to client/src/components/maps
Moved MapComponent function out of CarteHistorique
Removed a debug console.log
Added call to escapeHTMLChars
Removed a dummy testGeojson from source

carteInventaire
Moved from client/src/components to client/src/components/maps
Moved MapComponent function out of the main function
protected scale against absent inventaire and changed color scale
to be a standard color scale rather than something custom
Added more parameters to the MapComponent so it updates when
one of the display parameters changes

carteSecAnalyse
moved from client/src/components to client/src/components/maps
moved mapComponent out of the main function
added call to escapeHTMLChars to clean HTML symbols

CarteValidation
moved from client/src/components to client/src/components/maps
moved MapComponent out of the main function
Added call to escapeHTMLChars to swap protected characters

ControlAnaReg
Moved from client/src/components to client/src/components/menus
Moved MapComponent out of main loop and into separate function

AnalyseCartographiqueQuartiers
Moved from client/src/components to client/src/panels
Moved MapComponent function out of main function

client/pages changes
Due to the move, the imports in the pages have been changed in order to
refer to correct location

client/utils/escapeHTMLChars

Commonized function to use for replacing reserved chars in html
