+++ client/src/components/maps/carteEnsRegTerr.tsx
@@ -1,8 +1,9 @@
 import React,{useRef,useEffect} from 'react';
-import { CarteEnsRegTerrProps } from '../types/InterfaceTypes';
+import { CarteEnsRegTerrProps } from '../../types/InterfaceTypes';
 import { MapContainer, TileLayer,GeoJSON,useMap } from 'react-leaflet';
 import L from 'leaflet';
-import { utiliserContexte } from '../contexte/ContexteImmobilisation';
+import { utiliserContexte } from '../../contexte/ContexteImmobilisation';
+import { escapeHTMLChars } from '../../utils/escapeHTMLChars';
 
 const CarteEnsRegTerr:React.FC<CarteEnsRegTerrProps>=(props:CarteEnsRegTerrProps)=>{
 
@@ -15,7 +16,38 @@ const CarteEnsRegTerr:React.FC<CarteEnsRegTerrProps>=(props:CarteEnsRegTerrProps
     const attributionCarto = optionsCartos.find((entree)=>entree.id===optionCartoChoisie)?.attribution??'&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
     const zoomCarto = optionsCartos.find((entree)=>entree.id===optionCartoChoisie)?.zoomMax??18
     const geoJsonLayerGroupRef = useRef<L.LayerGroup | null>(null); // Refe
-    const MapComponent = () => {
+ 
+    
+    return(
+        <div className="div-carte-ens-reg-terr">
+        <MapContainer
+            center={props.centre}
+            zoom={props.zoom}
+            style={{ height: '100%', width: '100%' }}>
+            <TileLayer
+                url={urlCarto}
+                attribution={attributionCarto}
+                maxZoom={zoomCarto}
+                minZoom={1}
+                
+            />
+            <MapComponent
+              territoireSelect={props.territoireSelect}
+              geoJsonLayerGroupRef={geoJsonLayerGroupRef}
+            />
+        </MapContainer>
+        </div>
+    )
+
+}
+
+function MapComponent({ 
+  territoireSelect, 
+  geoJsonLayerGroupRef 
+}: { 
+  territoireSelect: CarteEnsRegTerrProps['territoireSelect']; 
+  geoJsonLayerGroupRef: React.RefObject<L.LayerGroup | null> 
+}){
             const map = useMap(); // Access the map instance
         
             useEffect(() => {
@@ -24,9 +56,9 @@ const CarteEnsRegTerr:React.FC<CarteEnsRegTerrProps>=(props:CarteEnsRegTerrProps
                   geoJsonLayerGroupRef.current.clearLayers(); // Clear previous vector layers
                 }
         
-                if (props.territoireSelect && props.territoireSelect.features && props.territoireSelect.features.length > 0) {
+                if (territoireSelect && territoireSelect.features && territoireSelect.features.length > 0) {
                   // Create a new GeoJSON layer from props.geoJsondata
-                  const geoJsonLayer = L.geoJSON(props.territoireSelect, {
+                  const geoJsonLayer = L.geoJSON(territoireSelect, {
                     style: {
                       color: 'blue', // Border color
                       weight: 2,     // Border thickness
@@ -37,9 +69,9 @@ const CarteEnsRegTerr:React.FC<CarteEnsRegTerrProps>=(props:CarteEnsRegTerrProps
                         if (feature.properties) {
                           const { id_periode_geo, secteur, ville } = feature.properties; // Destructure properties
                           const formattedPopupContent = `
-                            <strong>Feature ID:</strong> ${id_periode_geo} <br/>
-                            <strong>Name:</strong> ${ville} <br/>
-                            <strong>Secteur:</strong> ${secteur} <br/>
+                            <strong>Feature ID:</strong> ${escapeHTMLChars(id_periode_geo)} <br/>
+                            <strong>Name:</strong> ${escapeHTMLChars(ville)} <br/>
+                            <strong>Secteur:</strong> ${escapeHTMLChars(secteur)} <br/>
                           `;
                           layer.bindPopup(formattedPopupContent);
                         }
@@ -57,29 +89,9 @@ const CarteEnsRegTerr:React.FC<CarteEnsRegTerrProps>=(props:CarteEnsRegTerrProps
                   map.fitBounds(bounds);
                 }
               }
-            }, [props.territoireSelect, map]); // Dependency on props.geoJsondata and map
+            }, [territoireSelect, map]); // Dependency on props.geoJsondata and map
         
             return null; // No need to render anything for the map component itself
           };
-    
-    return(
-        <div className="div-carte-ens-reg-terr">
-        <MapContainer
-            center={props.centre}
-            zoom={props.zoom}
-            style={{ height: '100%', width: '100%' }}>
-            <TileLayer
-                url={urlCarto}
-                attribution={attributionCarto}
-                maxZoom={zoomCarto}
-                minZoom={1}
-                
-            />
-            <MapComponent/>
-        </MapContainer>
-        </div>
-    )
-
-}
 
 export default CarteEnsRegTerr;
