+++ client/src/pages/AnalyseReglements.tsx
@@ -4,7 +4,7 @@ import './analyseReglements.css'
 import GraphiqueReglements from '../components/charts/GraphiqueReglement';
 import { entete_ensembles_reglement_stationnement } from '../types/DataTypes';
 import { serviceEnsemblesReglements } from '../services';
-import ControlAnaReg from '../components/ControlAnaReg';
+import ControlAnaReg from '../components/menus/ControlAnaReg';
 
 const AnalyseReglements:React.FC = () =>{
     const [nGraphiques,defNGraphiques] = useState<number>(4);
